refactor: 교차 도메인 조율 진입점을 Facade에서 ApplicationService로 정리
- ReservationFacade를 roomescape.application.ReservationApplicationService로 이동 및 리네임
- 비즈니스 규칙을 강제하는 조율 책임이 순수 GoF Facade가 아니라 Service Layer(Application Service)임을 이름으로 명시
- 스테레오타입을 @Component에서 @Service로 변경하여 per-도메인 서비스와 통일
- 컨트롤러 4개의 의존 타입과 필드명을 reservationApplicationService로 갱신
- 단위 테스트를 unit.application 패키지로 이동하고 클래스명을 ReservationApplicationServiceMockTest로 변경
